Move UserNotificationCenterProtocol to it’s own file
